Add first shot at testcode form sender.py
The idea is to feed data into the sender_queue, provide a mock-up function for
the actual send process (send_out_email()) and call handle_next_mail() using
that mock-up function.

Instead of sending out an e-mail, the mock-up function puts the e-mail into
a message queue, from where the test code can retrieve it and analyze the
produced e-mail.

Signed-off-by: Andreas Platschek <andi.platschek@gmail.com>