doctests: fix tests, so they work as any user (also root) and in any case
"any case" in the topic refers to the dict -- if the result is checked, the
order of dict (as this typie is by definition unordered) may (and often will)
vary. Thus we can't check order. Instead we use a temporal variable tmp and
then compare if it is equal to the dictionary with the expected values.

Signed-off-by: Andreas Platschek <andi.platschek@gmail.com>